requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg[binary]>=3.2
python-dotenv>=1.0
pytest>=8.0
pytest-django>=4.8
ruff>=0.5
mypy>=1.10
bandit>=1.7
debugpy>=1.8